@@ -3,9 +3,9 @@
  * SPDX-License-Identifier: AGPL-3.0-or-later
  */
 
-import Vue from 'vue'
+import { createApp } from 'vue'
 import PersonalSettings from './components/PersonalSettings.vue'
-Vue.mixin({ methods: { t, n } })
 
-const View = Vue.extend(PersonalSettings)
-new View().$mount('#watsonx_prefs')
+const app = createApp(PersonalSettings)
+app.mixin({ methods: { t, n } })
+app.mount('#watsonx_prefs')
